refactor: decouples parsing and generation
