+++ src/parse-and-generate.spec.ts
@@ -2,14 +2,14 @@ import { equal } from "node:assert";
 import { readFileSync } from "node:fs";
 import { fileURLToPath } from "node:url";
 import { join } from "node:path";
-import { readAndConvert } from "./read-and-convert.js";
+import generate from "./parse-and-generate.js";
 
 const __dirname = fileURLToPath(new URL(".", import.meta.url));
 
 describe("reads and converts", () => {
   it("returns a CODEOWNERS as a string when passed file names of valid virtual code owners & virtual teams", () => {
     equal(
-      readAndConvert(
+      generate(
         join(__dirname, "__mocks__", "VIRTUAL-CODEOWNERS.txt"),
         join(__dirname, "__mocks__", "virtual-teams.yml")
       ),
